customer_complain/requirements.txt
google-adk>=1.17.0
google-cloud-aiplatform[adk,agent-engines]>=1.124.0
pillow>=12.0.0
python-dotenv>=1.2.1
toolbox-core>=0.5.0
requests>=2.32.5